Creation of basic unit classes
The text of a .cpp file that contains the classes for the game's combat units plus their basic attributes and files.
A test is included in the main function of this file. In the upcoming versions I plan to add an attack function and
a system that determines whether a unit is weak or strong against another.